Sc 53 pixel neighbourhood (#103)
* SC_53 spatial neighbors

* SC_53 update notebook

* SC_53 add to api

* SC_53 support for most frequent spatial neighbors

* SC_53 update docs

* SC_53 log message

* SC_53 docs

+++ src/harpy/utils/_aggregate.py
@@ -536,7 +536,7 @@ class RasterAggregator:
 
         sanity = da.all((~da.isnan(dask_array)).sum(axis=1) == 1)
         # da.nansum ignores np.nan added by _aggregate_custom_block
-        results = da.nansum(dask_array, axis=1).reshape(-1, features)
+        results = da.nansum(dask_array, axis=1, dtype=dtype).reshape(-1, features)
 
         sanity, results = dask.compute(*[sanity, results])
